More complete control over TMC2130 via JSON and in general. Other quintic cleanup.

+++ g2core/device/step_dir_driver/step_dir_driver.h
@@ -75,7 +75,7 @@ struct StepDirStepper final : Stepper  {
 
     bool canStep() override { return !_step.isNull(); };
 
-    void setMicrosteps(const uint8_t microsteps) override {
+    void setMicrosteps(const uint16_t microsteps) override {
         if (!_enable.isNull()) {
             switch (microsteps) {
                 case (1): {
